refactor(react-doctor): resolve maintainability findings (#32)
Clears all warning-level Maintainability findings except no-giant-component,
which is deferred to a follow-up (300–620 line component splits are a broad,
behavior-changing refactor best done separately).

Dead code:
- Delete unused pro-examples/ HeroUI reference examples (22 files)
- Delete dead src/pages/signup.tsx (invite-only app, no signup route) and
  src/features/sidebar/types.ts (unused IconSvgProps)
- Drop the HeroUI-template demo exports from sidebar-items.tsx and the now-dead
  team-avatar.tsx; keep only the live items/settingsItem
- Remove unused updateConversation() from chat-store-service.ts
- De-export file-local Markdown and ChartSkeleton helpers
- Remove unused deps @tauri-apps/plugin-opener, react-spreadsheet, usehooks-ts

Module-scope hoisting (prefer-module-scope-*):
- Hoist pure helpers/values out of component bodies: isDateUnavailable,
  singleKeySelect (chart-builder, pivot-builder), axisProps, the file-upload
  drag handlers (collapsed to preventDragDefault), deal-lookup today(),
  settings exportFormats

Component-file structure (only-export-components):
- Split useAuth/AuthContext into auth-context-value.ts (matches the existing
  *-context-value.ts convention)
- Move funderColor/CHART_COLORS to chart-colors.ts
- Move SidebarItemType enum + types to sidebar-types.ts

False positive:
- Ignore src/hero.ts in doctor.config.json: it is the HeroUI Tailwind v4 plugin
  entry loaded via `@plugin` in index.css, which the dead-code scan can't see
