+++ package.json
@@ -31,7 +31,6 @@
     "@tauri-apps/api": "~2.8.0",
     "@tauri-apps/plugin-dialog": "~2.4.0",
     "@tauri-apps/plugin-fs": "~2.4.2",
-    "@tauri-apps/plugin-opener": "~2.5.0",
     "@types/node": "^22.13.11",
     "@types/react-router-dom": "^5.3.3",
     "buffer": "^6.0.3",
@@ -40,10 +39,8 @@
     "react-dom": "^18.3.1",
     "react-markdown": "^10.1.0",
     "react-router-dom": "^7.9.1",
-    "react-spreadsheet": "^0.10.1",
     "recharts": "^2.15.1",
-    "remark-gfm": "^4.0.1",
-    "usehooks-ts": "^3.1.1"
+    "remark-gfm": "^4.0.1"
   },
   "overrides": {
     "@internationalized/date": "3.9.0"
