+++ src/components/auth/invite-user.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Button, Input, Card, CardBody, CardHeader, Select, SelectItem } from "@heroui/react";
-import { useAuth } from "@/contexts/auth-context";
+import { useAuth } from "@/contexts/auth-context-value";
 import { AuthService } from "@services/auth-service";
 
 // react-doctor-disable-next-line react-doctor/prefer-useReducer -- these hold largely independent concerns (form fields, per-field validation errors, async submit status) that change at different times, so a single reducer would not improve consistency
