--- src/features/sidebar/types.ts
@@ -1,5 +1,0 @@
-import type { SVGProps } from "react";
-
-export type IconSvgProps = SVGProps<SVGSVGElement> & {
-  size?: number;
-};
